actual_requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1758663848525/work
black @ file:///home/conda/feedstock_root/build_artifacts/black-recipe_1728503730172/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1756599151967/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1758716087526/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1758270398201/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coverage @ file:///home/conda/feedstock_root/build_artifacts/coverage_1758500757703/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1729286672586/work
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1733376776721/work
django-cors-headers==4.6.0
django-decouple==2.1
django-extensions==3.2.3
django-redis==5.4.0
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1733256735222/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
flake8 @ file:///home/conda/feedstock_root/build_artifacts/flake8_1733246708104/work
gunicorn==23.0.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1733223141826/work
isodate @ file:///home/conda/feedstock_root/build_artifacts/isodate_1733230734792/work
isort @ file:///home/conda/feedstock_root/build_artifacts/isort_1733236185843/work
mccabe @ file:///home/conda/feedstock_root/build_artifacts/mccabe_1733216466933/work
mypy @ file:///home/conda/feedstock_root/build_artifacts/mypy-split_1729643069211/work
mypy_extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1745776566517/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1730588037451/work/dist/numpy-2.1.3-cp312-cp312-linux_x86_64.whl#sha256=1e120dee91af211b94f044893ab48bdd16fef819ebe6bad4ba71f6a6925bdb9e
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1744430457695/work
pathspec @ file:///home/conda/feedstock_root/build_artifacts/pathspec_1733233363808/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1758169192777/work
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1733216196861/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pyflakes @ file:///home/conda/feedstock_root/build_artifacts/pyflakes_1733216066937/work
pymongo @ file:///home/conda/feedstock_root/build_artifacts/pymongo_1727858699199/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1758436411/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/pytest_1733217473289/work
pytest-cov @ file:///home/conda/feedstock_root/build_artifacts/pytest-cov_1733223023082/work
pytest-django @ file:///home/conda/feedstock_root/build_artifacts/pytest-django_1734959605658/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-decouple==3.8
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1733243180666/work
python-json-logger @ file:///croot/python-json-logger_1734370021104/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
rdflib @ file:///home/conda/feedstock_root/build_artifacts/rdflib_1734277910816/work
redis==6.4.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
sentry-sdk==2.18.0
setuptools==80.9.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
SPARQLWrapper @ file:///home/conda/feedstock_root/build_artifacts/sparqlwrapper_1669072955217/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1734007102193/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1753796677/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1733206856906/work
wheel==0.45.1
whitenoise==6.8.2
zstandard==0.25.0
